+++ cli/commands/destroy.ts
@@ -6,14 +6,13 @@ import { deleteAllBrokerSGs } from "../../aws/security-groups/deleteAllBrokerSGs
 import { deleteBrokerRole } from "../../aws/IAM/deleteBrokerRole";
 import { deleteRabbitoryRole } from "../../aws/IAM/deleteRabbitoryRole";
 import { runWithSpinner } from "../utils/spinner";
-import { promptUserForRegion } from "../utils/promptUserForRegion";
+import { promptUserForAvailabilityZone } from "../utils/promptUserForAvailabilityZone";
 import chalk from "chalk";
 
 
 export const destroy = async () => {
   try {
-    await promptUserForRegion();
-
+    await promptUserForAvailabilityZone();
     await runWithSpinner("Deleting DynamoDB Table...", () => deleteTable(), "Deleted DynamoDB Table");
     await runWithSpinner("Terminating Control Panel EC2 instance...", () => deleteControlPanel(), "Terminated EC2 instance");
     await runWithSpinner("Deleting RabbitMQ Broker Instances...", () => deleteAllBrokerInstances(), "Deleted RabbitMQ Broker Instances");
@@ -22,7 +21,7 @@ export const destroy = async () => {
     await runWithSpinner("Deleting RMQ Broker IAM role...", () => deleteBrokerRole(), "Deleted RMQ Broker IAM role");
     await runWithSpinner("Deleting Rabbitory IAM role...", () => deleteRabbitoryRole(), "Deleted Rabbitory IAM role");
 
-    console.log('\n\n')
+    console.log('\nSuccYour Rabbitory Control Panel, Rabbitmq Instances, and assoicated AWS resources \n')
   } catch (error) {
     console.error(chalk.redBright("\nRabbitory destruction failed\n"), error, "\n");
   }
