+++ cli/utils/promptUserForAvailabilityZone.ts
